Control Duplex and Skip options
There are some frequencies where we are supposed to listen, but not
transmit. Examples for that would be the US NOAA or some beacons.

Apart of that, there are some frequencies with little to no traffic.
I prefer to skip those while scanning and tune there only when I
want to.