Add weather sonde tracker and DMR spot
I got weather sonde tracker recently. It listens to the frequency of the
weather baloon and displays baloon's coordinates. Sometimes I might
find myself in a hunt for a baloon, but without my tracker. I still
want to be able to tune quickly to the standard frequency Bulgarian
baloons transmit to.

Another cool gadget that I recently got is a DMR hotspot.
It transforms my analogue radio to a digital one by
transmitting/receiving on a predefined sumplex frequency.